+++ nn_dataflow/tests/unit_test/test_partition_scheme.py
@@ -22,6 +22,7 @@ import collections
 import itertools
 import unittest
 
+from nn_dataflow.core import FmapPosition, FmapRange
 from nn_dataflow.core import Layer, ConvLayer, PoolingLayer
 from nn_dataflow.core import NodeRegion
 from nn_dataflow.core import ParallelEnum as pe
@@ -156,6 +157,43 @@ class TestPartitionScheme(unittest.TestCase):
                          self.ps2.dim(pe.OFMP, pe.BATP, pe.INPP)
                          * PhyDim2(1, 1))
 
+    def test_fmap_range(self):
+        ''' Get fmap_range. '''
+        fr1 = FmapRange(FmapPosition(b=0, n=0, h=0, w=0),
+                        FmapPosition(b=8, n=64, h=28, w=28))
+        # Small ranges.
+        fr2 = FmapRange(FmapPosition(b=0, n=0, h=0, w=0),
+                        FmapPosition(b=1, n=1, h=1, w=1))
+        # Irregular values.
+        fr3 = FmapRange(FmapPosition(b=2, n=4, h=2, w=6),
+                        FmapPosition(b=5, n=11, h=13, w=13))
+
+        ps = self.ps2
+
+        # No overlap.
+        for fr in [fr1, fr2, fr3]:
+            pfr_list = [ps.fmap_range(fr, pidx) for pidx in ps.gen_pidx()]
+            for idx, pfr in enumerate(pfr_list):
+                for jdx in range(idx):
+                    self.assertEqual(pfr_list[jdx].overlap_size(pfr), 0)
+
+        pidx = (PhyDim2(1, 0), PhyDim2(4, 3), PhyDim2(0, 2), PhyDim2(0, 0))
+
+        self.assertEqual(ps.fmap_range(fr1, pidx),
+                         FmapRange(FmapPosition(b=1, n=32, h=22, w=16),
+                                   FmapPosition(b=2, n=48, h=28, w=22)))
+        self.assertEqual(ps.fmap_range(fr2, pidx),
+                         FmapRange(FmapPosition(b=0, n=0, h=0, w=0),
+                                   FmapPosition(b=0, n=0, h=1, w=0)))
+        self.assertEqual(ps.fmap_range(fr3, pidx),
+                         FmapRange(FmapPosition(b=2, n=7, h=10, w=10),
+                                   FmapPosition(b=3, n=9, h=13, w=11)))
+
+    def test_is_appl2frng(self):
+        ''' Get is_applicable_to_fmap_range. '''
+        self.assertFalse(self.ps1.is_applicable_to_fmap_range())
+        self.assertTrue(self.ps2.is_applicable_to_fmap_range())
+
     def test_part_layer(self):
         ''' Get part_layer. '''
         batch_size = 16
@@ -204,11 +242,124 @@ class TestPartitionScheme(unittest.TestCase):
                 return self
             def ops_per_neuron(self):
                 return 0
+            @staticmethod
+            def data_loops():
+                return None
+
+        layer = _Layer(self.ps1.size(pe.OUTP), self.ps1.size(pe.OFMP))
+        self.assertEqual(layer.total_ops(), 0)
+        self.assertIsNone(_Layer.data_loops())
 
         with self.assertRaisesRegexp(TypeError, 'PartitionScheme: .*layer.*'):
-            _ = self.ps1.part_layer(_Layer(self.ps1.size(pe.OUTP),
-                                           self.ps1.size(pe.OFMP)),
-                                    self.ps1.size(pe.BATP))
+            _ = self.ps1.part_layer(layer, self.ps1.size(pe.BATP))
+
+    def test_projection(self):
+        ''' Get projection. '''
+
+        def _make_region(dim):
+            return NodeRegion(origin=PhyDim2(0, 0), dim=PhyDim2(*dim),
+                              type=NodeRegion.DATA)
+
+        # Shrink.
+        part = PartitionScheme(order=(pe.BATP, pe.INPP, pe.OUTP, pe.OFMP),
+                               pdims=((2, 3), (1, 5), (4, 4), (1, 1)))
+        proj_part = part.projection(_make_region((4, 30)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (4, 30))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 3))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 5))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (2, 2))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+        # Shrink multiple.
+        proj_part = part.projection(_make_region((2, 2)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (2, 2))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 1))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 2))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (1, 1))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+        # Shrink non-dividable.
+        proj_part = part.projection(_make_region((3, 54)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (2, 45))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 3))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 5))
+        # For height, 3 // 2 = 1.
+        # For width, 54 // 5 = 10, 10 // 3 = 3.
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (1, 3))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+        # Shrink with INPP.
+        part = PartitionScheme(order=(pe.BATP, pe.INPP, pe.OUTP, pe.OFMP),
+                               pdims=((2, 3), (1, 5), (4, 4), (4, 4)))
+        proj_part = part.projection(_make_region((4, 30)), appl2frng=True)
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (4, 30))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (2, 2))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+        proj_part = part.projection(_make_region((4, 30)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (4, 30))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (1, 1))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (2, 2))
+
+        # Shrink all.
+        proj_part = part.projection(_make_region((1, 1)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (1, 1))
+
+        # Extend.
+        part = PartitionScheme(order=(pe.INPP, pe.BATP, pe.OUTP, pe.OFMP),
+                               pdims=((2, 3), (1, 5), (1, 1), (1, 1)))
+        proj_part = part.projection(_make_region((4, 30)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (4, 30))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 3))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 5))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (2, 2))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+        # Extend non-dividable.
+        proj_part = part.projection(_make_region((5, 40)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (4, 30))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 3))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 5))
+        # For height, 5 // 2 = 2.
+        # For width, 40 // (3 * 5) == 2.
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (2, 2))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+        # Extend with INPP.
+        part = PartitionScheme(order=(pe.INPP, pe.BATP, pe.OUTP, pe.OFMP),
+                               pdims=((2, 3), (1, 5), (1, 1), (4, 4)))
+        proj_part = part.projection(_make_region((4, 30)), appl2frng=True)
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (4, 30))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 3))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 5))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (2, 2))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+        # Both shrink and extend.
+        part = PartitionScheme(order=(pe.BATP, pe.INPP, pe.OUTP, pe.OFMP),
+                               pdims=((2, 3), (1, 5), (4, 4), (1, 1)))
+        proj_part = part.projection(_make_region((16, 16)))
+        self.assertTupleEqual(proj_part.order, part.order)
+        self.assertTupleEqual(proj_part.dim(), (16, 15))
+        self.assertTupleEqual(proj_part.dim(pe.OUTP), (2, 3))
+        self.assertTupleEqual(proj_part.dim(pe.OFMP), (1, 5))
+        self.assertTupleEqual(proj_part.dim(pe.BATP), (8, 1))
+        self.assertTupleEqual(proj_part.dim(pe.INPP), (1, 1))
+
+    def test_projection_empty_region(self):
+        ''' Get projection with empty region. '''
+        with self.assertRaisesRegexp(ValueError, 'PartitionScheme: .*region.*'):
+            _ = self.ps1.projection(NodeRegion(origin=PhyDim2(0, 0),
+                                               dim=PhyDim2(0, 0),
+                                               type=NodeRegion.DATA))
 
     def test_repr(self):
         ''' __repr__. '''